--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -417,21 +417,6 @@ if(HAVE_ATTRIBUTE_ALIGNED)
 endif()
 
 #
-# check for _Thread_local() support in the compiler
-#
-check_c_source_compiles(
-    "_Thread_local int test;
-    int main(void) {
-        (void)test;
-        return 0;
-    }"
-    HAVE_THREAD_LOCAL
-)
-if(HAVE_THREAD_LOCAL)
-    add_definitions(-DHAVE_THREAD_LOCAL)
-endif()
-
-#
 # check for __builtin_ctz() support in the compiler
 #
 check_c_source_compiles(