Previous discussion in issue #1608.