@KungFuJesus 2023-12-12
I _think_ powerpc has 32 bit atomics but not necessarily 64 bit.  This is just a vague recollection, though.