@KungFuJesus 2023-12-12
https://en.cppreference.com/w/c/atomic/atomic_store

Requires C11, but a possible solution.