@phprus 2023-12-12
``functable`` is initialized by stub functions.
Stubs select platform-specific functions, write ``functable``, and call the selected platform-specific function to avoid recursion if weak-memory CPU reorder load and store. Stubs do not read ``functable``.

For this PR, the platform must provide single guarantee: Atomic store of pointer-size variable.