@KungFuJesus 2023-12-12
> @KungFuJesus Not 64bit, only pointer size atomic write is need (4 byte on 32bit platform or 8 byte on 64bit platform).
> 
> C11 atomic require "[atomic object types](https://en.cppreference.com/w/c/language/atomic)". For this reason I use the builtin functions of gcc and clang.

Right, it doesn't much matter on ppc32 if your pointer is 32 bit.  The C11 bits are portable, at least to C11 capable compilers, anyway.  The atomic_store macro approach here should do the job just fine, but it would involve wrapping it seven ways to sunday for other platforms. MSVC, GCC, and Clang covers a lot, but there's also Solaris Studio, ICC, XL, Open64, etc.  Admittedly obscure, but if we _could_ rely on C11 for those, it might be multiple birds with one stone.