@phprus 2023-12-12
> Admittedly obscure, but if we _could_ rely on C11 for those, it might be multiple birds with one stone.

Quick reply: No.
C11 atomic require special type. Using an C11 atomic pointer will require rewriting all versioned function calls to atomic_load the pointer and call the function at that pointer (overhead and unnecessary atomic ops).
For this PR we need **hardware** guarantees, not compiler guarantees.

If ppc32 guarantees 32-bit write atomicity and ppc64 guarantees 64-bit write atomicity, then PPC support can be added.