@KungFuJesus 2023-12-12
Oh sorry I followed the link and see what you're getting at.  We'd need to decorate everything with that atomic keyword, which would implicitly leverage the atomic loads.  

> If ppc32 guarantees 32-bit write atomicity and ppc64 guarantees 64-bit write atomicity, then PPC support can be added.

I _think_ that it does.  

> For this PR we need hardware guarantees, not compiler guarantees.

Right, but, with the compiler guarantees, the hardware atomicity would come with it.  The compiler should be doing the right thing under the hood (for most ISAs, nothing at all, for others, using whatever atomic store is available with cache aware compare and exchanges / mutexes if needed).  The GCC and MSVC paths you have in that macro do, at the cost of only supporting 3 compilers.  I agree though that we should be able to avoid the atomic load, the C11 approach might make that difficult.