@KungFuJesus 2023-12-12
I'm also wondering, however difficult, if we could write a test case that tries to induce a torn store to make sure.  I have SPARC, PowerPC, and ARM hardware I can test against for it.  

Unfortunately reliably producing a race with UB is probably impossible to do deterministically.