@phprus 2023-12-12
CI errors:
```
[2023-12-12T19:25:57.298Z] ['error'] There was an error running the uploader: Error uploading to [https://codecov.io:](https://codecov.io/) Error: There was an error fetching the storage URL during POST: 502 -
```