@phprus 2023-12-13
Converted to draft.
I need to fix one inaccuracy in macros.