@Dead2 2023-12-13
I did a few quick benchmarks just to make sure there are no surprises.

On x86-64 it shows a 0.3% average compression speedup, practically no change for decompression.
On Aarch64 it shows a 0.13% compression speedup, well inside margin of error. Also almost a 1% decompression speedup.

Since these speedups are marginal and the tests were not extensive, the most interesting part is that they are not slower and that they always seem to err on the side of faster. I'd say that bodes well for this PR.
I also did a couple variations to provoke different cache alignments, they too turned out to be similarly faster overall.

```
## Baseline 5af152a6a377b45619c8860e2a25c68fa2199ed9  x86-64
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.722/1.741/1.753/0.011        0.435/0.451/0.459/0.008       94,127,290
 2     35.518%      2.893/2.906/2.917/0.008        0.445/0.450/0.460/0.005       75,282,961
 3     33.882%      3.430/3.471/3.499/0.024        0.414/0.424/0.433/0.007       71,816,478
 4     33.174%      3.761/3.792/3.827/0.024        0.405/0.419/0.427/0.008       70,315,668
 5     32.660%      4.119/4.177/4.255/0.045        0.402/0.409/0.417/0.006       69,225,542
 6     32.508%      4.657/4.724/4.778/0.047        0.398/0.402/0.405/0.003       68,902,222
 7     32.255%      6.033/6.122/6.195/0.064        0.389/0.405/0.411/0.007       68,366,800
 8     32.167%      8.751/8.851/8.966/0.082        0.401/0.410/0.416/0.005       68,180,776
 9     31.887%   12.260/12.364/12.423/0.050        0.393/0.401/0.409/0.007       67,586,442

 avg1  34.273%                        5.350                          0.419
 tot                                385.190                         30.175      653,804,179

   text    data     bss     dec     hex filename
 145452    1384       8  146844   23d9c libz-ng.so.2

## PR #1609
  Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 1     44.409%      1.695/1.706/1.716/0.007        0.439/0.448/0.454/0.005       94,127,290
 2     35.518%      2.840/2.851/2.858/0.006        0.439/0.450/0.456/0.005       75,282,961
 3     33.882%      3.392/3.432/3.462/0.023        0.415/0.425/0.430/0.005       71,816,478
 4     33.174%      3.750/3.782/3.838/0.034        0.412/0.418/0.424/0.004       70,315,668
 5     32.660%      4.092/4.131/4.163/0.022        0.402/0.408/0.418/0.007       69,225,542
 6     32.508%      4.587/4.647/4.696/0.041        0.390/0.405/0.414/0.008       68,902,222
 7     32.255%      6.104/6.161/6.226/0.039        0.399/0.408/0.415/0.005       68,366,800
 8     32.167%      8.734/8.854/8.962/0.076        0.397/0.404/0.412/0.005       68,180,776
 9     31.887%   12.329/12.444/12.666/0.115        0.393/0.405/0.412/0.006       67,586,442

 avg1  34.273%                        5.334                          0.419
 tot                                384.062                         30.163      653,804,179

   text    data     bss     dec     hex filename
 145444    1576       8  147028   23e54 libz-ng.so.2

```

```
## Baseline 5af152a6a377b45619c8860e2a25c68fa2199ed9  aarch64
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.012/0.021/0.024/0.005        0.028/0.035/0.040/0.004       15,737,543
 1     54.165%      0.473/0.514/0.544/0.027        0.117/0.139/0.159/0.016        8,523,602
 2     43.877%      0.865/0.885/0.903/0.012        0.123/0.168/0.204/0.031        6,904,649
 3     42.389%      1.146/1.175/1.188/0.015        0.129/0.148/0.173/0.014        6,670,476
 4     41.647%      1.285/1.305/1.317/0.011        0.111/0.131/0.163/0.017        6,553,763
 5     41.217%      1.409/1.435/1.446/0.014        0.125/0.135/0.141/0.006        6,485,981
 6     41.037%      1.645/1.666/1.680/0.012        0.116/0.134/0.148/0.012        6,457,780
 7     40.778%      2.096/2.132/2.150/0.018        0.119/0.132/0.143/0.011        6,416,924
 8     40.704%      2.605/2.627/2.643/0.014        0.107/0.121/0.149/0.015        6,405,244
 9     40.409%      3.171/3.188/3.201/0.011        0.120/0.132/0.144/0.011        6,358,951

 avg1  48.623%                        1.495                          0.128
 avg2  54.026%                        1.661                          0.142
 tot                                119.582                         10.202       76,514,913

   text    data     bss     dec     hex filename
 119449    1616       8  121073   1d8f1 libz-ng.so.2

## PR #1609
 Level   Comp   Comptime min/avg/max/stddev  Decomptime min/avg/max/stddev  Compressed size
 0    100.008%      0.008/0.019/0.024/0.006        0.016/0.031/0.040/0.008       15,737,543
 1     54.165%      0.494/0.524/0.538/0.015        0.101/0.130/0.163/0.021        8,523,602
 2     43.877%      0.862/0.880/0.899/0.012        0.119/0.159/0.186/0.026        6,904,649
 3     42.389%      1.156/1.173/1.185/0.011        0.103/0.139/0.161/0.018        6,670,476
 4     41.647%      1.286/1.305/1.316/0.010        0.104/0.127/0.146/0.017        6,553,763
 5     41.217%      1.384/1.413/1.432/0.017        0.120/0.137/0.161/0.012        6,485,981
 6     41.037%      1.650/1.673/1.695/0.015        0.129/0.138/0.144/0.006        6,457,780
 7     40.778%      2.104/2.127/2.155/0.014        0.109/0.136/0.154/0.016        6,416,924
 8     40.704%      2.600/2.618/2.632/0.013        0.123/0.131/0.141/0.006        6,405,244
 9     40.409%      3.176/3.197/3.209/0.011        0.112/0.135/0.160/0.018        6,358,951

 avg1  48.623%                        1.493                          0.126
 avg2  54.026%                        1.659                          0.140
 tot                                119.428                         10.107       76,514,913

   text    data     bss     dec     hex filename
 118281    1544       8  119833   1d419 libz-ng.so.2

```